+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.39.3
+Add refresh token update capabilities to OAuthAuthenticator
+
+## 0.39.2
+Make step and cursor_granularity optional
+
 ## 0.39.1
 Improve connector builder error messages